Add Yuce AutoClicker project files
This commit includes the initial version of the Yuce AutoClicker project. The files include the source code for simulating mouse clicks based on user input (CPS and click type). The application is designed for Windows and comes with an easy-to-use interface. The project is now uploaded and ready for testing and further improvements.